requirements.txt
beautifulsoup4==4.6.3
Click==7.0
coloredlogs==10.0
humanfriendly==4.17
monotonic==1.5
pdfkit==0.6.1
pkg-resources==0.0.0
python-dotenv==0.9.1
